Create default DiffCallback implementation
